Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    float a,b;
-    scanf("%f %f", &a,&b);
+    int a,b;
+    scanf("%d %d", &a,&b);
     char n;
     scanf("%c", &n);
     switch(n)
